Most of the refactor
