+++ pkgs/ffigen/test/test_utils.dart
@@ -6,6 +6,7 @@ import 'dart:io';
 
 import 'package:ffigen/src/code_generator.dart';
 import 'package:ffigen/src/config_provider/config.dart';
+import 'package:ffigen/src/config_provider/yaml_config.dart';
 import 'package:ffigen/src/strings.dart' as strings;
 import 'package:logging/logging.dart';
 import 'package:package_config/package_config_types.dart';
@@ -138,7 +139,7 @@ void logToArray(List<String> logArr, Level level) {
 }
 
 Config testConfig(String yamlBody, {String? filename}) {
-  return Config.fromYaml(
+  return YamlConfig.fromYaml(
     yaml.loadYaml(yamlBody) as yaml.YamlMap,
     filename: filename,
     packageConfig: PackageConfig([
